move fileops to new file; update imports and namespaces
