+++ src/nicktionary.py
@@ -14,8 +14,8 @@ from logging.handlers import RotatingFileHandler
 
 import evaluate
 import help
-import interface
 import str_utils
+import fileops
 
 logger = logging.getLogger(__name__)
 
@@ -36,79 +36,6 @@ def config_logging():
     file_handler.setFormatter(formatter)
     logger.addHandler(file_handler)
 
-def resource_path(relative_path):
-    '''Get absolute path to resource, works for dev and for PyInstaller.
-    (copied from StackOverflow:
-    https://stackoverflow.com/questions/7674790/bundling-data-files-with-pyinstaller-onefile/13790741#13790741)
-    
-    '''
-    base_path = getattr(sys, '_MEIPASS', os.path.dirname(os.path.abspath(__file__)))
-    return os.path.join(base_path, relative_path)
-
-def check_for_wordle_file():
-    '''Aborts execution if solution list not found in this directory.'''
-    if not os.path.isfile(resource_path('wordle_list.txt')):
-        logger.critical('wordle_list.txt not found in this directory. Aborting.')
-        input(interface.WORDLE_LIST_NOT_FOUND)
-        sys.exit(1)
-
-def read_wordle(option, indate=''):
-    '''Reads solution word from given game mode from solution file.
-
-    Inputs:
-    option (str): One of 'today', 'random', 'date'.
-    indate (str) (required for 'date'): Specific date to play.
-
-    Raises: ValueError if option not in ['today', 'random', 'date'].
-
-    Returns: Solution word (str)
-    '''
-    if option == 'TODAY':
-        wordle_date = datetime.date.today().strftime('%b %d %Y')
-    
-    elif option == 'RANDOM':
-        start_date = datetime.date(2021, 6, 19) # first date in wordle_list
-        rand_days = random.choice(range(2314)) # number of unique dates in wordle_list
-        random_date = start_date + datetime.timedelta(days=rand_days)
-
-        wordle_date = random_date.strftime('%b %d %Y')
-
-    elif option == 'DATE':
-        year = int(indate[0:4])
-        month = int(indate[4:6])
-        day = int(indate[6:8])
-
-        wordle_date = datetime.date(year, month, day).strftime('%b %d %Y')
-
-    else:
-        raise ValueError(
-            f"Invalid option {option}, must be one of "
-            "'today','random', 'date'."
-        )
-
-    wordle = ''
-    
-    with open(resource_path('wordle_list.txt')) as f:
-        for line in f:
-            line_list = line.split()
-            filedate = str_utils.pad_str(line_list[0:3])
-            if filedate == wordle_date:
-                wordle = line_list[-1]
-                break
-
-    if not wordle:
-        print("\nToday's date not found. Choosing a random date.")
-        return read_wordle('random')
-
-    print(textwrap.dedent(
-        f"""
-        Playing Wordle from {wordle_date}.
-        ---------------------------------------------------------------------------"""
-        )
-    )
-
-    return wordle
-
 def play(wordle):
     '''Evaluates user guesses against solution word, prints progress to console.
 
@@ -128,7 +55,7 @@ def play(wordle):
     i = 0
     won = False
     while i < 6:
-        guess = interface.prompt(f'GUESS #{i+1}: ')
+        guess = str_utils.prompt(f'GUESS #{i+1}: ')
 
         if guess == 'QUIT':
             sys.exit(0)
@@ -179,15 +106,15 @@ def main():
 
     Returns: None.
     '''
-    check_for_wordle_file()
+    fileops.check_for_wordle_file()
     
-    print(interface.SPLASH)
+    print(str_utils.SPLASH)
     time.sleep(1.5)
 
     while True:
         try:
-            response = interface.prompt(
-                interface.MAIN_MENU,
+            response = str_utils.prompt(
+                str_utils.MAIN_MENU,
                 ['PLAY','RANDOM','DATE','HELP','QUIT']
             )
         except ValueError:
@@ -196,15 +123,15 @@ def main():
             continue
 
         if response == 'PLAY':
-            play(read_wordle('TODAY'))
+            play(fileops.read_wordle('TODAY'))
 
         elif response == 'RANDOM':
-            play(read_wordle('RANDOM'))
+            play(fileops.read_wordle('RANDOM'))
         
         elif response == 'DATE':
-            str_date = interface.prompt("Enter a date in the format YYYYMMDD: ")
+            str_date = str_utils.prompt("Enter a date in the format YYYYMMDD: ")
             try:
-                play(read_wordle('DATE',str_date))
+                play(fileops.read_wordle('DATE',str_date))
             except ValueError:
                 print('\nInvalid date. Try again!\n')
                 time.sleep(1)
